refactor: move split_top_level_maps to string_utils shared leaf module
The nesting-aware Erlang-map scanner was a private function in
type_checker::native_types. This moves it to semantic_analysis::string_utils
— the designated shared-kernel module for scanning utilities (alongside
split_top_level and split_generic_base, BT-3089) — so it is discoverable
and reusable by any future consumer without duplication.

No behaviour change; the function body is identical. Adds six unit tests
covering empty input, single map, multiple maps, nested maps, binary
literals with commas, and plain-brace nesting.

See docs/development/architecture-principles.md §6 (Shared-Leaf-Module Pattern).
